Split `arena-tag.ts` into separate files
