+++ src/data/abilities/ab-attrs/post-defend-apply-entry-hazard-tag-ab-attr.ts
@@ -1,6 +1,6 @@
 import { PostDefendAbAttr } from "#abilities/post-defend-ab-attr";
 import { globalScene } from "#app/global-scene";
-import type { EntryHazardTag } from "#data/arena-tag";
+import type { EntryHazardTag } from "#arena-tags/entry-hazard-tag";
 import type { ArenaTagType } from "#enums/arena-tag-type";
 import type { Pokemon } from "#field/pokemon";
 import type { Move } from "#moves/move";
