--- src/data/arena-tag.ts
@@ -1,1640 +1,0 @@
-import { applyAbAttrs } from "#abilities/apply-ab-attrs";
-import type { BlockNonDirectDamageAbAttr } from "#abilities/block-non-direct-damage-ab-attr";
-import type { InfiltratorAbAttr } from "#abilities/infiltrator-ab-attr";
-import type { ProtectStatAbAttr } from "#abilities/protect-stat-ab-attr";
-import { CommonBattleAnim } from "#animations/common-battle-anim";
-import { globalScene } from "#app/global-scene";
-import { getPokemonNameWithAffix } from "#app/messages";
-import type { SkyDropTag } from "#battler-tags/sky-drop-tag";
-import { SCREEN_DOUBLES_DMG_FACTOR, SCREEN_SINGLES_DMG_FACTOR } from "#constants/game-constants";
-import { allMoves } from "#data/data-lists";
-import { AbAttrFlag } from "#enums/ab-attr-flag";
-import { AbilityId } from "#enums/ability-id";
-import { ArenaTagSide } from "#enums/arena-tag-side";
-import { ArenaTagType } from "#enums/arena-tag-type";
-import type { BattlerIndex } from "#enums/battler-index";
-import { BattlerTagType } from "#enums/battler-tag-type";
-import { CommonAnim } from "#enums/common-anim";
-import { ElementalType } from "#enums/elemental-type";
-import { HitResult } from "#enums/hit-result";
-import { MoveCategory } from "#enums/move-category";
-import { MoveFlags } from "#enums/move-flags";
-import { MoveId } from "#enums/move-id";
-import { MoveTarget } from "#enums/move-target";
-import { PhaseId } from "#enums/phase-id";
-import { Stat } from "#enums/stat";
-import { StatusEffect } from "#enums/status-effect";
-import type { Arena } from "#field/arena";
-import type { Pokemon } from "#field/pokemon";
-import { PokemonMove } from "#field/pokemon-move";
-import { CommonAnimPhase } from "#phases/common-anim-phase";
-import { MoveEffectPhase } from "#phases/move-effect-phase";
-import { ShowAbilityPhase } from "#phases/show-ability-phase";
-import { StatStageChangePhase } from "#phases/stat-stage-change-phase";
-import { BooleanHolder, isNil, NumberHolder, toDmgValue } from "#utils/common-utils";
-import i18next from "i18next";
-
-export abstract class ArenaTag {
-  constructor(
-    public tagType: ArenaTagType,
-    public turnCount: number,
-    public sourceMoveId?: MoveId,
-    public sourceId?: number,
-    public side: ArenaTagSide = ArenaTagSide.BOTH,
-  ) {}
-
-  public get i18nSideKey(): string {
-    if (this.side === ArenaTagSide.PLAYER) {
-      return "Player";
-    }
-    if (this.side === ArenaTagSide.ENEMY) {
-      return "Enemy";
-    }
-    return "";
-  }
-
-  public apply(_arena: Arena, _simulated: boolean, ..._args: unknown[]): boolean {
-    return true;
-  }
-
-  public onAdd(_arena: Arena, _quiet: boolean = false): void {}
-
-  public onRemove(_arena: Arena, quiet: boolean = false): void {
-    if (!quiet) {
-      globalScene.phaseManager.queueMessagePhase(
-        i18next.t(`arenaTag:arenaOnRemove${this.i18nSideKey}`, { moveName: this.getMoveName() }),
-      );
-    }
-  }
-
-  public onOverlap(_arena: Arena): void {}
-
-  public lapse(_arena: Arena): boolean {
-    return this.turnCount < 1 || --this.turnCount !== 0;
-  }
-
-  public getMoveName(): string | null {
-    return this.sourceMoveId ? allMoves.get(this.sourceMoveId).name : null;
-  }
-
-  /**
-   * When given a arena tag or json representing one, load the data for it.
-   * This is meant to be inherited from by any arena tag with custom attributes
-   * @param source - The {@linkcode ArenaTag} source to load from
-   */
-  public loadTag(source: ArenaTag | any): void {
-    this.turnCount = source.turnCount;
-    this.sourceMoveId = source.sourceMoveId;
-    this.sourceId = source.sourceId;
-    this.side = source.side;
-  }
-
-  /**
-   * Helper function that retrieves the source Pokemon
-   * @returns The source {@linkcode Pokemon} or `null` if none is found
-   */
-  public getSourcePokemon(): Pokemon | null {
-    return this.sourceId ? globalScene.getPokemonById(this.sourceId) : null;
-  }
-
-  /**
-   * Helper function that retrieves the Pokemon affected
-   * @returns list of PlayerPokemon or EnemyPokemon on the field
-   */
-  public getAffectedPokemon(): Pokemon[] {
-    switch (this.side) {
-      case ArenaTagSide.PLAYER:
-        return globalScene.getPlayerField() ?? [];
-      case ArenaTagSide.ENEMY:
-        return globalScene.getEnemyField() ?? [];
-      case ArenaTagSide.BOTH:
-      default:
-        return globalScene.getField(true) ?? [];
-    }
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Mist_(move) Mist}.
- * Prevents Pokémon on the opposing side from lowering the stats of the Pokémon in the Mist.
- */
-export class MistTag extends ArenaTag {
-  constructor(turnCount: number, sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.MIST, turnCount, MoveId.MIST, sourceId, side);
-  }
-
-  override onAdd(arena: Arena, quiet: boolean = false): void {
-    super.onAdd(arena);
-
-    if (this.sourceId) {
-      const source = globalScene.getPokemonById(this.sourceId);
-
-      if (!quiet && source) {
-        globalScene.phaseManager.queueMessagePhase(
-          i18next.t("arenaTag:mistOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(source) }),
-        );
-      } else if (!quiet) {
-        console.warn("Failed to get source for MistTag onAdd");
-      }
-    }
-  }
-
-  /**
-   * Cancels the lowering of stats
-   * @param _arena the {@linkcode Arena} containing this effect
-   * @param simulated `true` if the effect should be applied quietly
-   * @param attacker the {@linkcode Pokemon} using a move into this effect.
-   * @param cancelled a {@linkcode BooleanHolder} whose value is set to `true`
-   * to flag the stat reduction as cancelled
-   * @returns `true` if a stat reduction was cancelled; `false` otherwise
-   */
-  override apply(_arena: Arena, simulated: boolean, attacker: Pokemon, cancelled: BooleanHolder): boolean {
-    if (attacker?.isActive(true)) {
-      const bypassed = new BooleanHolder(false);
-      applyAbAttrs<InfiltratorAbAttr>(AbAttrFlag.INFILTRATOR, attacker, simulated, bypassed);
-      if (bypassed.value) {
-        return false;
-      }
-    }
-
-    cancelled.value = true;
-
-    if (!simulated) {
-      globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:mistApply"));
-    }
-
-    return true;
-  }
-}
-
-/**
- * Reduces the damage of specific move categories in the arena.
- * @extends ArenaTag
- */
-export abstract class WeakenMoveScreenTag extends ArenaTag {
-  protected weakenedCategories: MoveCategory[];
-
-  /**
-   * Creates a new instance of the WeakenMoveScreenTag class.
-   *
-   * @param tagType - The type of the arena tag.
-   * @param turnCount - The number of turns the tag is active.
-   * @param sourceMoveId - The move that created the tag.
-   * @param sourceId - The ID of the source of the tag.
-   * @param side - The side (player or enemy) the tag affects.
-   * @param weakenedCategories - The categories of moves that are weakened by this tag.
-   */
-  constructor(
-    tagType: ArenaTagType,
-    turnCount: number,
-    sourceMoveId: MoveId,
-    sourceId: number,
-    side: ArenaTagSide,
-    weakenedCategories: MoveCategory[],
-  ) {
-    super(tagType, turnCount, sourceMoveId, sourceId, side);
-
-    this.weakenedCategories = weakenedCategories;
-  }
-
-  /**
-   * Applies the weakening effect to the move.
-   *
-   * @param _arena the {@linkcode Arena} where the move is applied.
-   * @param _simulated n/a
-   * @param attacker the attacking {@linkcode Pokemon}
-   * @param moveCategory the attacking move's {@linkcode MoveCategory}.
-   * @param damageMultiplier A {@linkcode NumberHolder} containing the damage multiplier
-   * @returns `true` if the attacking move was weakened; `false` otherwise.
-   */
-  override apply(
-    _arena: Arena,
-    simulated: boolean,
-    attacker: Pokemon,
-    moveCategory: MoveCategory,
-    damageMultiplier: NumberHolder,
-  ): boolean {
-    if (this.weakenedCategories.includes(moveCategory)) {
-      const bypassed = new BooleanHolder(false);
-      applyAbAttrs<InfiltratorAbAttr>(AbAttrFlag.INFILTRATOR, attacker, simulated, bypassed);
-      if (bypassed.value) {
-        return false;
-      }
-      damageMultiplier.value = globalScene.currentBattle.double ? SCREEN_DOUBLES_DMG_FACTOR : SCREEN_SINGLES_DMG_FACTOR;
-      return true;
-    }
-    return false;
-  }
-}
-
-/**
- * Reduces the damage of physical moves.
- * Used by {@linkcode MoveId.REFLECT}
- */
-class ReflectTag extends WeakenMoveScreenTag {
-  constructor(turnCount: number, sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.REFLECT, turnCount, MoveId.REFLECT, sourceId, side, [MoveCategory.PHYSICAL]);
-  }
-
-  override onAdd(_arena: Arena, quiet: boolean = false): void {
-    if (!quiet) {
-      globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:reflectOnAdd${this.i18nSideKey}`));
-    }
-  }
-}
-
-/**
- * Reduces the damage of special moves.
- * Used by {@linkcode MoveId.LIGHT_SCREEN}
- */
-class LightScreenTag extends WeakenMoveScreenTag {
-  constructor(turnCount: number, sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.LIGHT_SCREEN, turnCount, MoveId.LIGHT_SCREEN, sourceId, side, [MoveCategory.SPECIAL]);
-  }
-
-  override onAdd(_arena: Arena, quiet: boolean = false): void {
-    if (!quiet) {
-      globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:lightScreenOnAdd${this.i18nSideKey}`));
-    }
-  }
-}
-
-/**
- * Reduces the damage of physical and special moves.
- * Used by {@linkcode MoveId.AURORA_VEIL}
- */
-class AuroraVeilTag extends WeakenMoveScreenTag {
-  constructor(turnCount: number, sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.AURORA_VEIL, turnCount, MoveId.AURORA_VEIL, sourceId, side, [
-      MoveCategory.SPECIAL,
-      MoveCategory.PHYSICAL,
-    ]);
-  }
-
-  override onAdd(_arena: Arena, quiet: boolean = false): void {
-    if (!quiet) {
-      globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:auroraVeilOnAdd${this.i18nSideKey}`));
-    }
-  }
-}
-
-type ProtectConditionFunc = (arena: Arena, moveId: MoveId) => boolean;
-
-/**
- * Class to implement conditional team protection
- * applies protection based on the attributes of incoming moves
- */
-export abstract class ConditionalProtectTag extends ArenaTag {
-  /** The condition function to determine which moves are negated */
-  protected protectConditionFunc: ProtectConditionFunc;
-  /** Does this apply to all moves, including those that ignore other forms of protection? */
-  protected ignoresBypass: boolean;
-
-  constructor(
-    tagType: ArenaTagType,
-    sourceMoveId: MoveId,
-    sourceId: number,
-    side: ArenaTagSide,
-    condition: ProtectConditionFunc,
-    ignoresBypass: boolean = false,
-  ) {
-    super(tagType, 1, sourceMoveId, sourceId, side);
-
-    this.protectConditionFunc = condition;
-    this.ignoresBypass = ignoresBypass;
-  }
-
-  override onAdd(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(
-      i18next.t(`arenaTag:conditionalProtectOnAdd${this.i18nSideKey}`, { moveName: super.getMoveName() }),
-    );
-  }
-
-  // Removes default message for effect removal
-  override onRemove(_arena: Arena): void {}
-
-  /**
-   * Checks incoming moves against the condition function
-   * and protects the target if conditions are met
-   * @param arena the {@linkcode Arena} containing this tag
-   * @param simulated `true` if the tag is applied quietly; `false` otherwise.
-   * @param isProtected a {@linkcode BooleanHolder} used to flag if the move is protected against
-   * @param attacker the attacking {@linkcode Pokemon}
-   * @param defender the defending {@linkcode Pokemon}
-   * @param moveId the {@linkcode MoveId | identifier} for the move being used
-   * @param ignoresProtectBypass a {@linkcode BooleanHolder} used to flag if a protection effect supercedes effects that ignore protection
-   * @returns `true` if this tag protected against the attack; `false` otherwise
-   */
-  override apply(
-    arena: Arena,
-    simulated: boolean,
-    isProtected: BooleanHolder,
-    attacker: Pokemon,
-    defender: Pokemon,
-    moveId: MoveId,
-  ): boolean {
-    if (
-      (this.side === ArenaTagSide.PLAYER) === defender.isPlayer()
-      && this.protectConditionFunc(arena, moveId)
-      && (this.ignoresBypass || !allMoves.get(moveId).checkFlag(MoveFlags.IGNORE_PROTECT, attacker, defender))
-    ) {
-      if (!isProtected.value) {
-        isProtected.value = true;
-        if (!simulated) {
-          new CommonBattleAnim(CommonAnim.PROTECT, defender).play();
-          globalScene.phaseManager.queueMessagePhase(
-            i18next.t("arenaTag:conditionalProtectApply", {
-              moveName: super.getMoveName(),
-              pokemonNameWithAffix: getPokemonNameWithAffix(defender),
-            }),
-          );
-        }
-      }
-      return true;
-    }
-    return false;
-  }
-}
-
-/**
- * Condition function for {@link https://bulbapedia.bulbagarden.net/wiki/Quick_Guard_(move) Quick Guard's}
- * protection effect.
- * @param _arena {@linkcode Arena} The arena containing the protection effect
- * @param moveId {@linkcode MoveId} The move to check against this condition
- * @returns `true` if the incoming move's priority is greater than 0.
- *   This includes moves with modified priorities from abilities (e.g. Prankster)
- */
-const QuickGuardConditionFunc: ProtectConditionFunc = (_arena, moveId) => {
-  const move = allMoves.get(moveId);
-  const effectPhase = globalScene.phaseManager.getCurrentPhase();
-
-  if (effectPhase?.is<MoveEffectPhase>(PhaseId.MOVE_EFFECT)) {
-    const attacker = effectPhase.getUserPokemon();
-    if (attacker) {
-      return move.getPriority(attacker) > 0;
-    }
-  }
-  return move.priority > 0;
-};
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Quick_Guard_(move) Quick Guard}
- * Condition: The incoming move has increased priority.
- */
-class QuickGuardTag extends ConditionalProtectTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.QUICK_GUARD, MoveId.QUICK_GUARD, sourceId, side, QuickGuardConditionFunc);
-  }
-}
-
-/**
- * Condition function for {@link https://bulbapedia.bulbagarden.net/wiki/Wide_Guard_(move) Wide Guard's}
- * protection effect.
- * @param _arena {@linkcode Arena} The arena containing the protection effect
- * @param moveId {@linkcode MoveId} The move to check against this condition
- * @returns `true` if the incoming move is multi-targeted (even if it's only used against one Pokemon).
- */
-const WideGuardConditionFunc: ProtectConditionFunc = (_arena, moveId): boolean => {
-  const move = allMoves.get(moveId);
-
-  switch (move.moveTarget) {
-    case MoveTarget.ALL_ENEMIES:
-    case MoveTarget.ALL_NEAR_ENEMIES:
-    case MoveTarget.ALL_OTHERS:
-    case MoveTarget.ALL_NEAR_OTHERS:
-      return true;
-  }
-  return false;
-};
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Wide_Guard_(move) Wide Guard}
- * Condition: The incoming move can target multiple Pokemon. The move's source
- * can be an ally or enemy.
- */
-class WideGuardTag extends ConditionalProtectTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.WIDE_GUARD, MoveId.WIDE_GUARD, sourceId, side, WideGuardConditionFunc);
-  }
-}
-
-/**
- * Condition function for {@link https://bulbapedia.bulbagarden.net/wiki/Mat_Block_(move) Mat Block's}
- * protection effect.
- * @param _arena {@linkcode Arena} The arena containing the protection effect.
- * @param moveId {@linkcode MoveId} The move to check against this condition.
- * @returns `true` if the incoming move is not a Status move.
- */
-const MatBlockConditionFunc: ProtectConditionFunc = (_arena, moveId): boolean => {
-  const move = allMoves.get(moveId);
-  return move.category !== MoveCategory.STATUS;
-};
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Mat_Block_(move) Mat Block}
- * Condition: The incoming move is a Physical or Special attack move.
- */
-class MatBlockTag extends ConditionalProtectTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.MAT_BLOCK, MoveId.MAT_BLOCK, sourceId, side, MatBlockConditionFunc);
-  }
-
-  override onAdd(_arena: Arena) {
-    if (this.sourceId) {
-      const source = globalScene.getPokemonById(this.sourceId);
-      if (source) {
-        globalScene.phaseManager.queueMessagePhase(
-          i18next.t("arenaTag:matBlockOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(source) }),
-        );
-      } else {
-        console.warn("Failed to get source for MatBlockTag onAdd");
-      }
-    }
-  }
-}
-
-/**
- * Condition function for {@link https://bulbapedia.bulbagarden.net/wiki/Crafty_Shield_(move) Crafty Shield's}
- * protection effect.
- * @param _arena {@linkcode Arena} The arena containing the protection effect
- * @param moveId {@linkcode MoveId} The move to check against this condition
- * @returns `true` if the incoming move is a Status move, is not a hazard, and does not target all
- * Pokemon or sides of the field.
- */
-const CraftyShieldConditionFunc: ProtectConditionFunc = (_arena, moveId) => {
-  const move = allMoves.get(moveId);
-  return (
-    move.category === MoveCategory.STATUS
-    && move.moveTarget !== MoveTarget.ENEMY_SIDE
-    && move.moveTarget !== MoveTarget.BOTH_SIDES
-    && move.moveTarget !== MoveTarget.ALL
-  );
-};
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Crafty_Shield_(move) Crafty Shield}
- * Condition: The incoming move is a Status move, is not a hazard, and does
- * not target all Pokemon or sides of the field.
- */
-class CraftyShieldTag extends ConditionalProtectTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.CRAFTY_SHIELD, MoveId.CRAFTY_SHIELD, sourceId, side, CraftyShieldConditionFunc, true);
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Lucky_Chant_(move) Lucky Chant}.
- * Prevents critical hits against the tag's side.
- */
-export class NoCritTag extends ArenaTag {
-  /**
-   * Constructor method for the NoCritTag class
-   * @param turnCount `number` the number of turns this effect lasts
-   * @param sourceMoveId {@linkcode MoveId} the move that created this effect
-   * @param sourceId `number` the ID of the {@linkcode Pokemon} that created this effect
-   * @param side {@linkcode ArenaTagSide} the side to which this effect belongs
-   */
-  constructor(turnCount: number, sourceMoveId: MoveId, sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.NO_CRIT, turnCount, sourceMoveId, sourceId, side);
-  }
-
-  /** Queues a message upon adding this effect to the field */
-  override onAdd(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(
-      i18next.t(`arenaTag:noCritOnAdd${this.i18nSideKey}`, {
-        moveName: this.getMoveName(),
-      }),
-    );
-  }
-
-  /** Queues a message upon removing this effect from the field */
-  override onRemove(_arena: Arena): void {
-    const source = globalScene.getPokemonById(this.sourceId!); // TODO: is this bang correct?
-    globalScene.phaseManager.queueMessagePhase(
-      i18next.t("arenaTag:noCritOnRemove", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(source ?? undefined),
-        moveName: this.getMoveName(),
-      }),
-    );
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Wish_(move) Wish}.
- * Heals the Pokémon in the user's position the turn after Wish is used.
- */
-class WishTag extends ArenaTag {
-  private battlerIndex: BattlerIndex;
-  private triggerMessage: string;
-  private healHp: number;
-
-  constructor(turnCount: number, sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.WISH, turnCount, MoveId.WISH, sourceId, side);
-  }
-
-  override onAdd(_arena: Arena): void {
-    if (this.sourceId) {
-      const user = globalScene.getPokemonById(this.sourceId);
-      if (user) {
-        this.battlerIndex = user.getBattlerIndex();
-        this.triggerMessage = i18next.t("arenaTag:wishTagOnAdd", {
-          pokemonNameWithAffix: getPokemonNameWithAffix(user),
-        });
-        this.healHp = toDmgValue(user.getMaxHp() / 2);
-      } else {
-        console.warn("Failed to get source for WishTag onAdd");
-      }
-    }
-  }
-
-  override onRemove(_arena: Arena): void {
-    const target = globalScene.getPokemonByBattlerIndex(this.battlerIndex);
-    if (target?.isActive(true)) {
-      globalScene.phaseManager.queueMessagePhase(this.triggerMessage);
-      globalScene.phaseManager.queuePokemonHealPhase(target.getBattlerIndex(), this.healHp);
-    }
-  }
-}
-
-/**
- * Abstract class to implement weakened moves of a specific type.
- */
-export abstract class WeakenMoveTypeTag extends ArenaTag {
-  private weakenedType: ElementalType;
-
-  /**
-   * Creates a new instance of the WeakenMoveTypeTag class.
-   *
-   * @param tagType - The type of the arena tag.
-   * @param turnCount - The number of turns the tag is active.
-   * @param type - The type being weakened from this tag.
-   * @param sourceMoveId - The move that created the tag.
-   * @param sourceId - The ID of the source of the tag.
-   */
-  constructor(tagType: ArenaTagType, turnCount: number, type: ElementalType, sourceMoveId: MoveId, sourceId: number) {
-    super(tagType, turnCount, sourceMoveId, sourceId);
-
-    this.weakenedType = type;
-  }
-
-  /**
-   * Reduces an attack's power by 0.33x if it matches this tag's weakened type.
-   * @param _arena n/a
-   * @param _simulated n/a
-   * @param type the attack's {@linkcode ElementalType}
-   * @param power a {@linkcode NumberHolder} containing the attack's power
-   * @returns `true` if the attack's power was reduced; `false` otherwise.
-   */
-  override apply(_arena: Arena, _simulated: boolean, type: ElementalType, power: NumberHolder): boolean {
-    if (type === this.weakenedType) {
-      power.value *= 0.33;
-      return true;
-    }
-    return false;
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Mud_Sport_(move) Mud Sport}.
- * Weakens Electric type moves for a set amount of turns, usually 5.
- */
-class MudSportTag extends WeakenMoveTypeTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(ArenaTagType.MUD_SPORT, turnCount, ElementalType.ELECTRIC, MoveId.MUD_SPORT, sourceId);
-  }
-
-  override onAdd(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:mudSportOnAdd"));
-  }
-
-  override onRemove(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:mudSportOnRemove"));
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Water_Sport_(move) Water Sport}.
- * Weakens Fire type moves for a set amount of turns, usually 5.
- */
-class WaterSportTag extends WeakenMoveTypeTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(ArenaTagType.WATER_SPORT, turnCount, ElementalType.FIRE, MoveId.WATER_SPORT, sourceId);
-  }
-
-  override onAdd(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:waterSportOnAdd"));
-  }
-
-  override onRemove(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:waterSportOnRemove"));
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Ion_Deluge_(move) | Ion Deluge}
- * and the secondary effect of {@link https://bulbapedia.bulbagarden.net/wiki/Plasma_Fists_(move) | Plasma Fists}.
- * Converts Normal-type moves to Electric type for the rest of the turn.
- */
-export class IonDelugeTag extends ArenaTag {
-  constructor(sourceMoveId?: MoveId) {
-    super(ArenaTagType.ION_DELUGE, 1, sourceMoveId);
-  }
-
-  /** Queues an on-add message */
-  override onAdd(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:plasmaFistsOnAdd"));
-  }
-
-  override onRemove(_arena: Arena): void {} // Removes default on-remove message
-
-  /**
-   * Converts Normal-type moves to Electric type
-   * @param _arena n/a
-   * @param _simulated n/a
-   * @param moveType a {@linkcode NumberHolder} containing a move's {@linkcode ElementalType}
-   * @returns `true` if the given move type changed; `false` otherwise.
-   */
-  override apply(_arena: Arena, _simulated: boolean, moveType: NumberHolder): boolean {
-    if (moveType.value === ElementalType.NORMAL) {
-      moveType.value = ElementalType.ELECTRIC;
-      return true;
-    }
-    return false;
-  }
-}
-
-/**
- * Abstract class to implement arena entry hazards.
- * @extends ArenaTag
- */
-export abstract class EntryHazardTag extends ArenaTag {
-  public layers: number;
-  public maxLayers: number;
-
-  /**
-   * @param tagType - The type of the arena tag.
-   * @param sourceMoveId - The move that created the tag.
-   * @param sourceId - The ID of the source of the tag.
-   * @param side - The side (player or enemy) the tag affects.
-   * @param maxLayers - The maximum amount of layers this tag can have.
-   */
-  constructor(tagType: ArenaTagType, sourceMoveId: MoveId, sourceId: number, side: ArenaTagSide, maxLayers: number) {
-    super(tagType, 0, sourceMoveId, sourceId, side);
-
-    this.layers = 1;
-    this.maxLayers = maxLayers;
-  }
-
-  override onOverlap(arena: Arena): void {
-    if (this.layers < this.maxLayers) {
-      this.layers++;
-
-      this.onAdd(arena);
-    }
-  }
-
-  /**
-   * Activates the hazard effect onto a Pokemon when it enters the field
-   * @param _arena the {@linkcode Arena} containing this tag
-   * @param simulated if `true`, only checks if the hazard would activate.
-   * @param pokemon the {@linkcode Pokemon} triggering this hazard
-   * @returns `true` if this hazard affects the given Pokemon; `false` otherwise.
-   */
-  override apply(_arena: Arena, simulated: boolean, pokemon: Pokemon): boolean {
-    if (this.side !== ArenaTagSide.BOTH && (this.side === ArenaTagSide.PLAYER) !== pokemon.isPlayer()) {
-      return false;
-    }
-
-    return this.activateTrap(pokemon, simulated);
-  }
-
-  activateTrap(_pokemon: Pokemon, _simulated: boolean): boolean {
-    return false;
-  }
-
-  getMatchupScoreMultiplier(pokemon: Pokemon): number {
-    return pokemon.isGrounded()
-      ? 1
-      : Phaser.Math.Linear(0, 1 / Math.pow(2, this.layers), Math.min(pokemon.getHpRatio(), 0.5) * 2);
-  }
-
-  override loadTag(source: any): void {
-    super.loadTag(source);
-    this.layers = source.layers;
-    this.maxLayers = source.maxLayers;
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Spikes_(move) Spikes}.
- * Applies up to 3 layers of Spikes, dealing 1/8th, 1/6th, or 1/4th of the the Pokémon's HP
- * in damage for 1, 2, or 3 layers of Spikes respectively if they are summoned into this trap.
- */
-class SpikesTag extends EntryHazardTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.SPIKES, MoveId.SPIKES, sourceId, side, 3);
-  }
-
-  override onAdd(arena: Arena, quiet: boolean = false): void {
-    super.onAdd(arena);
-
-    const source = this.sourceId ? globalScene.getPokemonById(this.sourceId) : null;
-    if (!quiet && source) {
-      globalScene.phaseManager.queueMessagePhase(
-        i18next.t("arenaTag:spikesOnAdd", {
-          moveName: this.getMoveName(),
-          opponentDesc: source.getOpponentDescriptor(),
-        }),
-      );
-    }
-  }
-
-  override activateTrap(pokemon: Pokemon, simulated: boolean): boolean {
-    if (pokemon.isGrounded()) {
-      const cancelled = new BooleanHolder(false);
-      applyAbAttrs<BlockNonDirectDamageAbAttr>(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, simulated, cancelled);
-
-      if (simulated) {
-        return !cancelled.value;
-      }
-
-      if (!cancelled.value) {
-        const damageHpRatio = 1 / (10 - 2 * this.layers);
-        const damage = toDmgValue(pokemon.getMaxHp() * damageHpRatio);
-
-        globalScene.phaseManager.queueMessagePhase(
-          i18next.t("arenaTag:spikesActivateTrap", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-        );
-        pokemon.damageAndUpdate(damage, { result: HitResult.OTHER });
-        return true;
-      }
-    }
-
-    return false;
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Toxic_Spikes_(move) Toxic Spikes}.
- * Applies up to 2 layers of Toxic Spikes, poisoning or badly poisoning any Pokémon who is
- * summoned into this trap if 1 or 2 layers of Toxic Spikes respectively are up. Poison-type
- * Pokémon summoned into this trap remove it entirely.
- */
-class ToxicSpikesTag extends EntryHazardTag {
-  private neutralized: boolean;
-
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.TOXIC_SPIKES, MoveId.TOXIC_SPIKES, sourceId, side, 2);
-    this.neutralized = false;
-  }
-
-  override onAdd(arena: Arena, quiet: boolean = false): void {
-    super.onAdd(arena);
-
-    const source = this.sourceId ? globalScene.getPokemonById(this.sourceId) : null;
-    if (!quiet && source) {
-      globalScene.phaseManager.queueMessagePhase(
-        i18next.t("arenaTag:toxicSpikesOnAdd", {
-          moveName: this.getMoveName(),
-          opponentDesc: source.getOpponentDescriptor(),
-        }),
-      );
-    }
-  }
-
-  override onRemove(arena: Arena): void {
-    if (!this.neutralized) {
-      super.onRemove(arena);
-    }
-  }
-
-  override activateTrap(pokemon: Pokemon, simulated: boolean): boolean {
-    if (pokemon.isGrounded()) {
-      if (simulated) {
-        return true;
-      }
-      if (pokemon.isOfType(ElementalType.POISON)) {
-        this.neutralized = true;
-        if (globalScene.arena.removeTag(this.tagType)) {
-          globalScene.phaseManager.queueMessagePhase(
-            i18next.t("arenaTag:toxicSpikesActivateTrapPoison", {
-              pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-              moveName: this.getMoveName(),
-            }),
-          );
-          return true;
-        }
-      } else if (!pokemon.hasNonVolatileStatusEffect()) {
-        const inflictsToxic = this.layers > 1;
-        return pokemon.trySetStatus(
-          inflictsToxic ? StatusEffect.TOXIC : StatusEffect.POISON,
-          true,
-          null,
-          0,
-          this.getMoveName(),
-        );
-      }
-    }
-
-    return false;
-  }
-
-  override getMatchupScoreMultiplier(pokemon: Pokemon): number {
-    if (pokemon.isGrounded() || !pokemon.canSetStatus(StatusEffect.POISON, true)) {
-      return 1;
-    }
-    if (pokemon.isOfType(ElementalType.POISON)) {
-      return 1.25;
-    }
-    return super.getMatchupScoreMultiplier(pokemon);
-  }
-}
-
-/**
- * Interface representing a delayed attack command.
- * @see {@linkcode DelayedAttackTag}
- */
-interface DelayedAttack {
-  sourceId: number;
-  moveId: MoveId;
-  targetIndex: BattlerIndex;
-  turnCount: number;
-}
-
-/**
- * Arena Tag class for delayed attacks from {@link https://bulbapedia.bulbagarden.net/wiki/Future_Sight_(move) Future Sight}
- * and {@link https://bulbapedia.bulbagarden.net/wiki/Doom_Desire_(move) Doom Desire}.
- * Delays the attack's effect by 3 turns (including the turn the move is used),
- * and deals damage after the turn count is reached.
- */
-export class DelayedAttackTag extends ArenaTag {
-  /** Contains all queued delayed attacks on the field */
-  public delayedAttacks: DelayedAttack[];
-
-  constructor() {
-    super(ArenaTagType.DELAYED_ATTACK, 0);
-
-    this.delayedAttacks = [];
-  }
-
-  public addAttack(source: Pokemon, moveId: MoveId, targetIndex: BattlerIndex): void {
-    this.delayedAttacks.push({ sourceId: source.id, moveId: moveId, targetIndex, turnCount: 3 });
-  }
-
-  override lapse(_arena: Arena): boolean {
-    this.delayedAttacks.forEach((attack) => {
-      attack.turnCount--;
-
-      if (!isNil(globalScene.getPokemonById(attack.sourceId)) && attack.turnCount <= 0) {
-        const target = globalScene.getField(true).find((p) => attack.targetIndex === p.getBattlerIndex());
-        if (target) {
-          globalScene.phaseManager.unshiftPhase(
-            new MoveEffectPhase(attack.sourceId, [attack.targetIndex], new PokemonMove(attack.moveId, 0, 0, true)),
-          );
-        } else if (globalScene.currentBattle.double) {
-          const redirectIndex = attack.targetIndex + (attack.targetIndex % 2 === 0 ? 1 : -1);
-          globalScene.phaseManager.unshiftPhase(
-            new MoveEffectPhase(attack.sourceId, [redirectIndex], new PokemonMove(attack.moveId, 0, 0, true)),
-          );
-        }
-      }
-    });
-
-    this.delayedAttacks = this.delayedAttacks.filter(
-      (attack) => !isNil(globalScene.getPokemonById(attack.sourceId)) && attack.turnCount > 0,
-    );
-    return this.delayedAttacks.length > 0;
-  }
-
-  override onRemove(_arena: Arena): void {}
-
-  override loadTag(source: ArenaTag | any): void {
-    super.loadTag(source);
-    this.delayedAttacks = source.delayedAttacks;
-  }
-}
-
-/**
- * Contains data related to a queued healing effect from
- * {@link https://bulbapedia.bulbagarden.net/wiki/Healing_Wish_(move) | Healing Wish}
- * or {@link https://bulbapedia.bulbagarden.net/wiki/Lunar_Dance_(move) | Lunar Dance}.
- */
-interface PendingHealEffect {
-  /** The id for the {@linkcode Pokemon} that created the effect */
-  readonly sourceId: number;
-  /** The {@linkcode MoveId | id} for the move that created the effect */
-  readonly moveId: MoveId;
-  /** If `true`, also restores the target's PP when the effect activates */
-  readonly restorePP: boolean;
-  /** The `i18n` key for the message to display when the effect activates */
-  readonly healMessageKey: string;
-}
-
-/**
- * Arena tag to contain stored healing effects, namely from
- * {@link https://bulbapedia.bulbagarden.net/wiki/Healing_Wish_(move) | Healing Wish}
- * and {@link https://bulbapedia.bulbagarden.net/wiki/Lunar_Dance_(move) | Lunar Dance}.
- * When a damaged Pokemon first enters the effect's {@linkcode BattlerIndex | field position},
- * their HP is fully restored, and they are cured of any non-volatile status condition.
- * If the effect is from Lunar Dance, their PP is also restored.
- * @extends ArenaTag
- */
-export class PendingHealTag extends ArenaTag {
-  /** All pending healing effects, organized by {@linkcode BattlerIndex} */
-  private pendingHeals: Partial<Record<BattlerIndex, PendingHealEffect[]>> = {};
-
-  constructor() {
-    super(ArenaTagType.PENDING_HEAL, 0);
-  }
-
-  /**
-   * Adds a pending healing effect to the field. Effects under the same move *and*
-   * target index as an existing effect are ignored.
-   * @param targetIndex - The {@linkcode BattlerIndex} under which the effect applies
-   * @param healEffect - The {@linkcode PendingHealEffect | data} for the pending heal effect
-   */
-  public queueHeal(targetIndex: BattlerIndex, healEffect: PendingHealEffect): void {
-    const existingHealEffects = this.pendingHeals[targetIndex];
-    if (existingHealEffects) {
-      if (!existingHealEffects.some((he) => he.moveId === healEffect.moveId)) {
-        existingHealEffects.push(healEffect);
-      }
-    } else {
-      this.pendingHeals[targetIndex] = [healEffect];
-    }
-  }
-
-  /** Removes default on-remove message */
-  override onRemove(_arena: Arena): void {}
-
-  /** This arena tag is removed at the end of the turn if no pending healing effects are on the field */
-  override lapse(_arena: Arena): boolean {
-    for (const key in this.pendingHeals) {
-      if (this.pendingHeals[key].length > 0) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  /**
-   * Applies a pending healing effect on the given target index. If an effect is found for
-   * the index, the Pokemon at that index is healed to full HP, is cured of any non-volatile status,
-   * and has its PP fully restored (if the effect is from Lunar Dance).
-   * @param arena - The {@linkcode Arena} containing this tag
-   * @param simulated - If `true`, suppresses changes to game state
-   * @param pokemon - The {@linkcode Pokemon} receiving the healing effect
-   * @returns `true` if the target Pokemon was healed by this effect
-   * @todo This should also be called when a Pokemon moves into a new position via Ally Switch
-   */
-  override apply(arena: Arena, simulated: boolean, pokemon: Pokemon): boolean {
-    const targetIndex = pokemon.getBattlerIndex();
-    const targetEffects = this.pendingHeals[targetIndex];
-
-    if (simulated) {
-      return !!targetEffects?.length;
-    }
-
-    const healEffect = targetEffects?.find((effect) => this.canApply(effect, pokemon));
-    if (targetEffects && healEffect) {
-      const { sourceId, moveId, restorePP, healMessageKey } = healEffect;
-      const sourcePokemon = globalScene.getPokemonById(sourceId);
-      if (!sourcePokemon) {
-        console.warn(`Source of pending ${allMoves.get(moveId).name} effect is undefined!`);
-        targetEffects.splice(targetEffects.indexOf(healEffect), 1);
-        // Re-evaluate after the invalid heal effect is removed
-        return this.apply(arena, simulated, pokemon);
-      }
-
-      globalScene.phaseManager.queuePokemonHealPhase(targetIndex, pokemon.getMaxHp(), {
-        message: i18next.t(healMessageKey, { pokemonName: getPokemonNameWithAffix(sourcePokemon) }),
-        healStatus: true,
-        fullRestorePP: restorePP,
-      });
-
-      targetEffects.splice(targetEffects.indexOf(healEffect), 1);
-    }
-
-    return !isNil(healEffect);
-  }
-
-  /**
-   * Determines if the given {@linkcode PendingHealEffect} can immediately heal
-   * the given target {@linkcode Pokemon}.
-   * @param healEffect - The {@linkcode PendingHealEffect} to evaluate
-   * @param pokemon - The {@linkcode Pokemon} to evaluate against
-   * @returns `true` if the Pokemon can be healed by the effect
-   */
-  private canApply(healEffect: PendingHealEffect, pokemon: Pokemon): boolean {
-    return (
-      !pokemon.isFullHp()
-      || pokemon.hasNonVolatileStatusEffect()
-      || (healEffect.restorePP && pokemon.getMoveset().some((mv) => mv.ppUsed > 0))
-    );
-  }
-
-  override loadTag(source: ArenaTag | any): void {
-    super.loadTag(source);
-    this.pendingHeals = source.pendingHeals;
-  }
-}
-
-/**
- * Class used for hazards that damage based on type. The two existing ones are
- * Stealth rock (produced by stealth rock and stone axe) and
- * Sharp steel (produced by G-Max steelsurge)
- */
-class TypeHazardTag extends EntryHazardTag {
-  public readonly damagingType: ElementalType;
-  public readonly onAddKey: string;
-  public readonly activateTrapKey: string;
-
-  constructor(
-    arenaTagType: ArenaTagType,
-    damagingType: ElementalType,
-    sourceId: number,
-    side: ArenaTagSide,
-    sourceMoveId: MoveId,
-    onAddKey: string,
-    activateTrapKey: string,
-  ) {
-    super(arenaTagType, sourceMoveId, sourceId, side, 1);
-    this.damagingType = damagingType;
-    this.onAddKey = onAddKey;
-    this.activateTrapKey = activateTrapKey;
-  }
-
-  override onAdd(arena: Arena, quiet: boolean = false): void {
-    super.onAdd(arena);
-
-    const source = this.sourceId ? globalScene.getPokemonById(this.sourceId) : null;
-    if (!quiet && source) {
-      globalScene.phaseManager.queueMessagePhase(
-        i18next.t(this.onAddKey, { opponentDesc: source.getOpponentDescriptor() }),
-      );
-    }
-  }
-
-  getDamageHpRatio(pokemon: Pokemon): number {
-    const effectiveness = pokemon.getAttackTypeEffectiveness(this.damagingType, undefined, true);
-    return effectiveness * 0.125;
-  }
-
-  override activateTrap(pokemon: Pokemon, simulated: boolean): boolean {
-    const cancelled = new BooleanHolder(false);
-    applyAbAttrs<BlockNonDirectDamageAbAttr>(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, simulated, cancelled);
-
-    if (cancelled.value) {
-      return false;
-    }
-
-    const damageHpRatio = this.getDamageHpRatio(pokemon);
-
-    if (damageHpRatio) {
-      if (simulated) {
-        return true;
-      }
-      const damage = toDmgValue(pokemon.getMaxHp() * damageHpRatio);
-      globalScene.phaseManager.queueMessagePhase(
-        i18next.t(this.activateTrapKey, { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-      pokemon.damageAndUpdate(damage, { result: HitResult.OTHER });
-      return true;
-    }
-
-    return false;
-  }
-
-  override getMatchupScoreMultiplier(pokemon: Pokemon): number {
-    const damageHpRatio = this.getDamageHpRatio(pokemon);
-    return Phaser.Math.Linear(super.getMatchupScoreMultiplier(pokemon), 1, 1 - Math.pow(damageHpRatio, damageHpRatio));
-  }
-}
-
-class StealthRockTag extends TypeHazardTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(
-      ArenaTagType.STEALTH_ROCK,
-      ElementalType.ROCK,
-      sourceId,
-      side,
-      MoveId.STEALTH_ROCK,
-      "arenaTag:stealthRockOnAdd",
-      "arenaTag:stealthRockActivateTrap",
-    );
-  }
-}
-
-class SharpSteelTag extends TypeHazardTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(
-      ArenaTagType.SHARP_STEEL,
-      ElementalType.STEEL,
-      sourceId,
-      side,
-      MoveId.G_MAX_STEELSURGE,
-      "arenaTag:sharpSteelOnAdd",
-      "arenaTag:sharpSteelActivateTrap",
-    );
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Sticky_Web_(move) Sticky Web}.
- * Applies up to 1 layer of Sticky Web, which lowers the Speed by one stage
- * to any Pokémon who is summoned into this trap.
- */
-class StickyWebTag extends EntryHazardTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.STICKY_WEB, MoveId.STICKY_WEB, sourceId, side, 1);
-  }
-
-  /** @todo Should `quiet` ever be `true`? */
-  override onAdd(arena: Arena, quiet: boolean = false): void {
-    super.onAdd(arena);
-    const source = this.sourceId ? globalScene.getPokemonById(this.sourceId) : null;
-    if (!quiet && source) {
-      globalScene.phaseManager.queueMessagePhase(
-        i18next.t(`arenaTag:stickyWebOnAdd${this.i18nSideKey}Side`, {
-          moveName: this.getMoveName(),
-          opponentDesc: source.getOpponentDescriptor(),
-        }),
-      );
-    }
-  }
-
-  override activateTrap(pokemon: Pokemon, simulated: boolean): boolean {
-    if (pokemon.isGrounded()) {
-      const cancelled = new BooleanHolder(false);
-      applyAbAttrs<ProtectStatAbAttr>(AbAttrFlag.PROTECT_STAT, pokemon, simulated, Stat.SPD, cancelled);
-
-      if (simulated) {
-        return !cancelled.value;
-      }
-
-      if (!cancelled.value) {
-        globalScene.phaseManager.queueMessagePhase(
-          i18next.t("arenaTag:stickyWebActivateTrap", { pokemonName: pokemon.getNameToRender() }),
-        );
-        const stages = new NumberHolder(-1);
-        globalScene.phaseManager.unshiftPhase(
-          new StatStageChangePhase(pokemon.getBattlerIndex(), this.getSourcePokemon(), [Stat.SPD], stages.value, {
-            isStickyWeb: true,
-          }),
-        );
-        return true;
-      }
-    }
-
-    return false;
-  }
-}
-
-/**
- * Base class for moves like Trick Room which should negate their effect when used a second time.
- */
-export abstract class ArenaRoomTag extends ArenaTag {
-  override onOverlap(arena: Arena): void {
-    arena.removeTag(this.tagType);
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Trick_Room_(move) Trick Room}.
- * Reverses the Speed calculation for all Pokémon on the field as long as this arena tag is up.
- */
-export class TrickRoomTag extends ArenaRoomTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(ArenaTagType.TRICK_ROOM, turnCount, MoveId.TRICK_ROOM, sourceId);
-  }
-
-  /**
-   * @param speedReversed - A {@linkcode BooleanHolder} used to flag if Speed-based
-   * turn order should be reversed.
-   * @returns `true`
-   */
-  override apply(_arena: Arena, _simulated: boolean, speedReversed: BooleanHolder): boolean {
-    speedReversed.value = !speedReversed.value;
-    return true;
-  }
-
-  override onAdd(_arena: Arena): void {
-    const source = this.sourceId ? globalScene.getPokemonById(this.sourceId) : null;
-    if (source) {
-      globalScene.phaseManager.queueMessagePhase(
-        i18next.t("arenaTag:trickRoomOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(source) }),
-      );
-    }
-  }
-
-  override onRemove(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:trickRoomOnRemove"));
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Gravity_(move) Gravity}.
- * Grounds all Pokémon on the field, including Flying-types and those with
- * {@linkcode AbilityId.LEVITATE} for the duration of the arena tag, usually 5 turns.
- */
-export class GravityTag extends ArenaTag {
-  constructor(turnCount: number) {
-    super(ArenaTagType.GRAVITY, turnCount, MoveId.GRAVITY);
-  }
-
-  override onAdd(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:gravityOnAdd"));
-    globalScene.getField(true).forEach((pokemon) => {
-      if (pokemon) {
-        pokemon.removeTag(BattlerTagType.FLOATING);
-        pokemon.removeTag(BattlerTagType.TELEKINESIS);
-        if (pokemon.getTag(BattlerTagType.FLYING)) {
-          pokemon.addTag(BattlerTagType.INTERRUPTED);
-        }
-        pokemon.getTag<SkyDropTag>(BattlerTagType.SKY_DROP)?.clearSkyDropEffects();
-      }
-    });
-  }
-
-  override onRemove(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:gravityOnRemove"));
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Tailwind_(move) Tailwind}.
- * Doubles the Speed of the Pokémon who created this arena tag, as well as all allied Pokémon.
- * Applies this arena tag for 4 turns (including the turn the move was used).
- */
-class TailwindTag extends ArenaTag {
-  constructor(turnCount: number, sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.TAILWIND, turnCount, MoveId.TAILWIND, sourceId, side);
-  }
-
-  override onAdd(_arena: Arena, quiet: boolean = false): void {
-    if (!quiet) {
-      globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:tailwindOnAdd${this.i18nSideKey}`));
-    }
-
-    const source = globalScene.getPokemonById(this.sourceId!); //TODO: this bang is questionable!
-    const party = source?.getField() ?? [];
-
-    for (const pokemon of party) {
-      // Apply the CHARGED tag to party members with the WIND_POWER ability
-      if (pokemon.hasAbility(AbilityId.WIND_POWER) && !pokemon.getTag(BattlerTagType.CHARGED)) {
-        pokemon.addTag(BattlerTagType.CHARGED);
-        globalScene.phaseManager.queueMessagePhase(
-          i18next.t("abilityTriggers:windPowerCharged", {
-            pokemonName: getPokemonNameWithAffix(pokemon),
-            moveName: this.getMoveName(),
-          }),
-        );
-      }
-      // Raise attack by one stage if party member has WIND_RIDER ability
-      if (pokemon.hasAbility(AbilityId.WIND_RIDER)) {
-        globalScene.phaseManager.unshiftPhase(new ShowAbilityPhase(pokemon.getBattlerIndex()));
-        globalScene.phaseManager.unshiftPhase(
-          new StatStageChangePhase(pokemon.getBattlerIndex(), pokemon, [Stat.ATK], 1),
-        );
-      }
-    }
-  }
-
-  override onRemove(_arena: Arena, quiet: boolean = false): void {
-    if (!quiet) {
-      globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:tailwindOnRemove${this.i18nSideKey}`));
-    }
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Happy_Hour_(move) Happy Hour}.
- * Doubles the prize money from trainers and money moves like {@linkcode MoveId.PAY_DAY} and {@linkcode MoveId.MAKE_IT_RAIN}.
- */
-class HappyHourTag extends ArenaTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.HAPPY_HOUR, 0, MoveId.HAPPY_HOUR, sourceId, side);
-  }
-
-  override onAdd(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:happyHourOnAdd"));
-  }
-
-  override onRemove(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:happyHourOnRemove"));
-  }
-}
-
-class SafeguardTag extends ArenaTag {
-  constructor(turnCount: number, sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.SAFEGUARD, turnCount, MoveId.SAFEGUARD, sourceId, side);
-  }
-
-  override onAdd(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:safeguardOnAdd${this.i18nSideKey}`));
-  }
-
-  override onRemove(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:safeguardOnRemove${this.i18nSideKey}`));
-  }
-}
-
-class NoneTag extends ArenaTag {
-  constructor() {
-    super(ArenaTagType.NONE, 0);
-  }
-}
-
-/**
- * Arena Tag implementing the "sea of fire" effect from the combination
- * of {@link https://bulbapedia.bulbagarden.net/wiki/Fire_Pledge_(move) | Fire Pledge}
- * and {@link https://bulbapedia.bulbagarden.net/wiki/Grass_Pledge_(move) | Grass Pledge}.
- * Damages all non-Fire-type Pokemon on the given side of the field at the end
- * of each turn for 4 turns.
- */
-class FireGrassPledgeTag extends ArenaTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.FIRE_GRASS_PLEDGE, 4, MoveId.FIRE_PLEDGE, sourceId, side);
-  }
-
-  override onAdd(_arena: Arena): void {
-    // "A sea of fire enveloped your/the opposing team!"
-    globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:fireGrassPledgeOnAdd${this.i18nSideKey}`));
-  }
-
-  override lapse(arena: Arena): boolean {
-    const field: Pokemon[] =
-      this.side === ArenaTagSide.PLAYER ? globalScene.getPlayerField() : globalScene.getEnemyField();
-
-    field
-      .filter((pokemon) => pokemon.isActive(true) && !pokemon.isOfType(ElementalType.FIRE) && !pokemon.switchOutStatus)
-      .forEach((pokemon) => {
-        const cancelled = new BooleanHolder(false);
-        applyAbAttrs<BlockNonDirectDamageAbAttr>(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, false, cancelled);
-        if (cancelled.value) {
-          return;
-        }
-
-        // "{pokemonNameWithAffix} was hurt by the sea of fire!"
-        globalScene.phaseManager.queueMessagePhase(
-          i18next.t("arenaTag:fireGrassPledgeLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-        );
-        // TODO: Replace this with a proper animation
-        globalScene.phaseManager.unshiftPhase(
-          new CommonAnimPhase(CommonAnim.MAGMA_STORM, pokemon.getBattlerIndex(), pokemon.getBattlerIndex()),
-        );
-        pokemon.damageAndUpdate(toDmgValue(pokemon.getMaxHp() / 8));
-      });
-
-    return super.lapse(arena);
-  }
-}
-
-/**
- * Arena Tag implementing the "rainbow" effect from the combination
- * of {@link https://bulbapedia.bulbagarden.net/wiki/Water_Pledge_(move) | Water Pledge}
- * and {@link https://bulbapedia.bulbagarden.net/wiki/Fire_Pledge_(move) | Fire Pledge}.
- * Doubles the secondary effect chance of moves from Pokemon on the
- * given side of the field for 4 turns.
- */
-class WaterFirePledgeTag extends ArenaTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.WATER_FIRE_PLEDGE, 4, MoveId.WATER_PLEDGE, sourceId, side);
-  }
-
-  override onAdd(_arena: Arena): void {
-    // "A rainbow appeared in the sky on your/the opposing team's side!"
-    globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:waterFirePledgeOnAdd${this.i18nSideKey}`));
-  }
-
-  /**
-   * Doubles the chance for the given move's secondary effect(s) to trigger
-   * @param _arena the {@linkcode Arena} containing this tag
-   * @param _simulated n/a
-   * @param moveChance a {@linkcode NumberHolder} containing
-   * the move's current effect chance
-   * @returns `true` if the move's effect chance was doubled (currently always `true`)
-   */
-  override apply(_arena: Arena, _simulated: boolean, moveChance: NumberHolder): boolean {
-    moveChance.value *= 2;
-    return true;
-  }
-}
-
-/**
- * Arena Tag implementing the "swamp" effect from the combination
- * of {@link https://bulbapedia.bulbagarden.net/wiki/Grass_Pledge_(move) | Grass Pledge}
- * and {@link https://bulbapedia.bulbagarden.net/wiki/Water_Pledge_(move) | Water Pledge}.
- * Quarters the Speed of Pokemon on the given side of the field for 4 turns.
- */
-class GrassWaterPledgeTag extends ArenaTag {
-  constructor(sourceId: number, side: ArenaTagSide) {
-    super(ArenaTagType.GRASS_WATER_PLEDGE, 4, MoveId.GRASS_PLEDGE, sourceId, side);
-  }
-
-  override onAdd(_arena: Arena): void {
-    // "A swamp enveloped your/the opposing team!"
-    globalScene.phaseManager.queueMessagePhase(i18next.t(`arenaTag:grassWaterPledgeOnAdd${this.i18nSideKey}`));
-  }
-}
-
-/**
- * Class to describe the field effect of G-Max moves that damage all Pokemon that
- * are not immune by 1/6th of their health
- *
- * Used in:
- * G-Max Vine Lash: Grass
- * G-Max Wildfire: Fire
- * G-Max Cannonade: Water
- * G-Max Volcalith: Rock
- */
-export class TypeImmuneDamageOverTimeTag extends ArenaTag {
-  private immuneType: ElementalType;
-
-  constructor(tagType, sourceMoveId: MoveId, sourceId: number, side: ArenaTagSide, immuneType: ElementalType) {
-    super(tagType, 4, sourceMoveId, sourceId, side);
-    this.immuneType = immuneType;
-  }
-
-  private getAnimationForType() {
-    switch (this.immuneType) {
-      case ElementalType.GRASS:
-        return CommonAnim.WRAP;
-      case ElementalType.FIRE:
-        return CommonAnim.FIRE_SPIN;
-      case ElementalType.WATER:
-        return CommonAnim.WHIRLPOOL;
-      case ElementalType.ROCK:
-        return CommonAnim.SALT_CURE;
-      default:
-        return CommonAnim.WRAP;
-    }
-  }
-
-  override onAdd(_arena: Arena) {
-    globalScene.phaseManager.queueMessagePhase(
-      i18next.t(`arenaTag:TypeImmuneDamageOverTimeOnAdd${this.i18nSideKey}${ElementalType[this.immuneType]}`),
-    );
-  }
-
-  override lapse(arena: Arena): boolean {
-    const field: Pokemon[] =
-      this.side === ArenaTagSide.PLAYER ? globalScene.getPlayerField() : globalScene.getEnemyField();
-
-    field
-      .filter((pokemon) => pokemon.isActive(true) && !pokemon.isOfType(this.immuneType) && !pokemon.switchOutStatus)
-      .forEach((pokemon) => {
-        const cancelled = new BooleanHolder(false);
-        applyAbAttrs<BlockNonDirectDamageAbAttr>(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, false, cancelled);
-        if (cancelled.value) {
-          return;
-        }
-
-        globalScene.phaseManager.queueMessagePhase(
-          i18next.t(`arenaTag:TypeImmuneDamageOverTimeLapse${ElementalType[this.immuneType]}`, {
-            pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-          }),
-        );
-        // TODO: Replace this with a proper animation
-        globalScene.phaseManager.unshiftPhase(
-          new CommonAnimPhase(this.getAnimationForType(), pokemon.getBattlerIndex(), pokemon.getBattlerIndex()),
-        );
-        pokemon.damageAndUpdate(toDmgValue(pokemon.getMaxHp() / 6));
-      });
-
-    return super.lapse(arena);
-  }
-}
-
-/**
- * Arena Tag class for {@link https://bulbapedia.bulbagarden.net/wiki/Fairy_Lock_(move) Fairy Lock}.
- * Fairy Lock prevents all Pokémon (except Ghost types) on the field from switching out or
- * fleeing during their next turn.
- * If a Pokémon that's on the field when Fairy Lock is used goes on to faint later in the same turn,
- * the Pokémon that replaces it will still be unable to switch out in the following turn.
- */
-export class FairyLockTag extends ArenaTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(ArenaTagType.FAIRY_LOCK, turnCount, MoveId.FAIRY_LOCK, sourceId);
-  }
-
-  override onAdd(_arena: Arena): void {
-    globalScene.phaseManager.queueMessagePhase(i18next.t("arenaTag:fairyLockOnAdd"));
-  }
-}
-
-// TODO: swap `sourceMoveId` and `sourceId` and make `sourceMoveId` an optional parameter
-export function getArenaTag(
-  tagType: ArenaTagType,
-  sourceId: number,
-  turnCount: number,
-  sourceMoveId?: MoveId,
-  side: ArenaTagSide = ArenaTagSide.BOTH,
-): ArenaTag | null {
-  switch (tagType) {
-    case ArenaTagType.MIST:
-      return new MistTag(turnCount, sourceId, side);
-    case ArenaTagType.QUICK_GUARD:
-      return new QuickGuardTag(sourceId, side);
-    case ArenaTagType.WIDE_GUARD:
-      return new WideGuardTag(sourceId, side);
-    case ArenaTagType.MAT_BLOCK:
-      return new MatBlockTag(sourceId, side);
-    case ArenaTagType.CRAFTY_SHIELD:
-      return new CraftyShieldTag(sourceId, side);
-    case ArenaTagType.NO_CRIT:
-      return new NoCritTag(turnCount, sourceMoveId!, sourceId, side); // TODO: is this bang correct?
-    case ArenaTagType.MUD_SPORT:
-      return new MudSportTag(turnCount, sourceId);
-    case ArenaTagType.WATER_SPORT:
-      return new WaterSportTag(turnCount, sourceId);
-    case ArenaTagType.ION_DELUGE:
-      return new IonDelugeTag(sourceMoveId);
-    case ArenaTagType.SPIKES:
-      return new SpikesTag(sourceId, side);
-    case ArenaTagType.TOXIC_SPIKES:
-      return new ToxicSpikesTag(sourceId, side);
-    case ArenaTagType.DELAYED_ATTACK:
-      return new DelayedAttackTag();
-    case ArenaTagType.PENDING_HEAL:
-      return new PendingHealTag();
-    case ArenaTagType.WISH:
-      return new WishTag(turnCount, sourceId, side);
-    case ArenaTagType.STEALTH_ROCK:
-      return new StealthRockTag(sourceId, side);
-    case ArenaTagType.STICKY_WEB:
-      return new StickyWebTag(sourceId, side);
-    case ArenaTagType.TRICK_ROOM:
-      return new TrickRoomTag(turnCount, sourceId);
-    case ArenaTagType.GRAVITY:
-      return new GravityTag(turnCount);
-    case ArenaTagType.REFLECT:
-      return new ReflectTag(turnCount, sourceId, side);
-    case ArenaTagType.LIGHT_SCREEN:
-      return new LightScreenTag(turnCount, sourceId, side);
-    case ArenaTagType.AURORA_VEIL:
-      return new AuroraVeilTag(turnCount, sourceId, side);
-    case ArenaTagType.TAILWIND:
-      return new TailwindTag(turnCount, sourceId, side);
-    case ArenaTagType.HAPPY_HOUR:
-      return new HappyHourTag(sourceId, side);
-    case ArenaTagType.SAFEGUARD:
-      return new SafeguardTag(turnCount, sourceId, side);
-    case ArenaTagType.FIRE_GRASS_PLEDGE:
-      return new FireGrassPledgeTag(sourceId, side);
-    case ArenaTagType.WATER_FIRE_PLEDGE:
-      return new WaterFirePledgeTag(sourceId, side);
-    case ArenaTagType.GRASS_WATER_PLEDGE:
-      return new GrassWaterPledgeTag(sourceId, side);
-    case ArenaTagType.FAIRY_LOCK:
-      return new FairyLockTag(turnCount, sourceId);
-    case ArenaTagType.G_MAX_VINE_LASH:
-      return new TypeImmuneDamageOverTimeTag(
-        ArenaTagType.G_MAX_VINE_LASH,
-        MoveId.G_MAX_VINE_LASH,
-        sourceId,
-        side,
-        ElementalType.GRASS,
-      );
-    case ArenaTagType.G_MAX_WILDFIRE:
-      return new TypeImmuneDamageOverTimeTag(
-        ArenaTagType.G_MAX_WILDFIRE,
-        MoveId.G_MAX_WILDFIRE,
-        sourceId,
-        side,
-        ElementalType.FIRE,
-      );
-    case ArenaTagType.G_MAX_CANNONADE:
-      return new TypeImmuneDamageOverTimeTag(
-        ArenaTagType.G_MAX_CANNONADE,
-        MoveId.G_MAX_CANNONADE,
-        sourceId,
-        side,
-        ElementalType.WATER,
-      );
-    case ArenaTagType.G_MAX_VOLCALITH:
-      return new TypeImmuneDamageOverTimeTag(
-        ArenaTagType.G_MAX_VOLCALITH,
-        MoveId.G_MAX_VOLCALITH,
-        sourceId,
-        side,
-        ElementalType.ROCK,
-      );
-    case ArenaTagType.SHARP_STEEL:
-      return new SharpSteelTag(sourceId, side);
-    default:
-      return null;
-  }
-}
-
-/**
- * When given a battler tag or json representing one, creates an actual ArenaTag object with the same data.
- * @param source - The source {@linkcode ArenaTag}
- * @returns The valid {@linkcode ArenaTag}
- */
-export function loadArenaTag(source: ArenaTag | any): ArenaTag {
-  const tag =
-    getArenaTag(source.tagType, source.sourceId, source.turnCount, source.sourceMoveId, source.side) ?? new NoneTag();
-  tag.loadTag(source);
-  return tag;
-}
