+++ src/data/arena-tags/arena-room-tag.ts
@@ -1,0 +1,11 @@
+import { ArenaTag } from "#arena-tags/arena-tag";
+import type { Arena } from "#field/arena";
+
+/**
+ * Base class for moves like Trick Room which should negate their effect when used a second time.
+ */
+export abstract class ArenaRoomTag extends ArenaTag {
+  override onOverlap(arena: Arena): void {
+    arena.removeTag(this.tagType);
+  }
+}
