+++ src/data/arena-tags/sharp-steel-tag.ts
@@ -1,0 +1,19 @@
+import { TypeHazardTag } from "#arena-tags/type-hazard-tag";
+import type { ArenaTagSide } from "#enums/arena-tag-side";
+import { ArenaTagType } from "#enums/arena-tag-type";
+import { ElementalType } from "#enums/elemental-type";
+import { MoveId } from "#enums/move-id";
+
+export class SharpSteelTag extends TypeHazardTag {
+  constructor(sourceId: number, side: ArenaTagSide) {
+    super(
+      ArenaTagType.SHARP_STEEL,
+      ElementalType.STEEL,
+      sourceId,
+      side,
+      MoveId.G_MAX_STEELSURGE,
+      "arenaTag:sharpSteelOnAdd",
+      "arenaTag:sharpSteelActivateTrap",
+    );
+  }
+}
