+++ src/data/moves/move-attrs/delayed-attack-attr.ts
@@ -1,6 +1,6 @@
 import { globalScene } from "#app/global-scene";
 import { getPokemonNameWithAffix } from "#app/messages";
-import type { DelayedAttackTag } from "#data/arena-tag";
+import type { DelayedAttackTag } from "#arena-tags/delayed-attack-tag";
 import { ArenaTagType } from "#enums/arena-tag-type";
 import type { ChargeAnim } from "#enums/charge-anim";
 import { MoveResult } from "#enums/move-result";
